requirements.txt
numpy==1.26.4
Pillow==10.3.0
psutil==5.9.8
scipy==1.13.1
skimage==0.22.0
